requirements.txt
aiohttp
asyncio
discord.py
pyyaml
tinydb
python-dateutil
emojis
humanize
tweepy
requests
nums_from_string
multiprocess